+++ bench/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bottom-bench"
+version = "0.1.0"
+authors = []
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+bench = false
+
+[[bench]]
+name = "bottom"
+harness = false
+path = "src/bench.rs"
+
+[dependencies]
+criterion = "0.3.3"
+bottomify = { version = "*", path = ".." }
