+++ src/bottom.rs
@@ -1,33 +1,7 @@
-use std::collections::HashMap;
-use lazy_static::lazy_static;
-use maplit::hashmap;
 use std::error::Error;
 use std::fmt;
 
-lazy_static! {
-    static ref CHARACTER_VALUES: HashMap<u8, &'static str> = hashmap! {
-        200 => "🫂" ,
-        50 => "💖",
-        10 => "✨",
-        5 => "🥺",
-        1 => ",",
-        0 => "❤️",
-    };
-    static ref BYTE_TO_EMOJI: HashMap<u8, String> = {
-        let mut m = HashMap::new();
-        for i in 0..=255 {
-            m.insert(i, byte_to_emoji(i));
-        }
-        m
-    };
-    static ref EMOJI_TO_BYTE: HashMap<&'static str, &'static u8> = {
-        let mut m = HashMap::new();
-        for (byte, emoji) in &mut BYTE_TO_EMOJI.iter() {
-            m.insert(emoji.as_str().trim_end_matches("👉👈"), byte);
-        }
-        m
-    };
-}
+include!(concat!(env!("OUT_DIR"), "/maps.rs"));
 
 #[derive(Debug)]
 pub struct TranslationError {
@@ -42,41 +16,8 @@ impl fmt::Display for TranslationError {
 
 impl Error for TranslationError {}
 
-fn byte_to_emoji(value: u8) -> String {
-    let mut buffer = String::new();
-    let mut value = value;
-
-    if value == 0 {
-        return CHARACTER_VALUES[&0].to_string();
-    }
-
-    loop {
-        let (to_push, subtract_by) = {
-            if value >= 200 {
-                (CHARACTER_VALUES[&200], 200)
-            } else if value >= 50 {
-                (CHARACTER_VALUES[&50], 50)
-            } else if value >= 10 {
-                (CHARACTER_VALUES[&10], 10)
-            } else if value >= 5 {
-                (CHARACTER_VALUES[&5], 5)
-            } else if value >= 1 {
-                (CHARACTER_VALUES[&1], 1)
-            } else {
-                break;
-            }
-        };
-
-        buffer.push_str(to_push);
-        value -= subtract_by;
-    }
-
-    buffer.push_str("👉👈");
-    buffer
-}
-
 pub fn encode_byte(value: u8) -> &'static str {
-    &BYTE_TO_EMOJI[&value]
+    &BYTE_TO_EMOJI[value as usize]
 }
 
 pub fn decode_byte(input: &dyn AsRef<str>) -> Result<u8, TranslationError> {
@@ -84,7 +25,7 @@ pub fn decode_byte(input: &dyn AsRef<str>) -> Result<u8, TranslationError> {
     let result = EMOJI_TO_BYTE.get(input_ref).ok_or(TranslationError {
         why: format!("Cannot decode character {}", input_ref),
     })?;
-    Ok(**result)
+    Ok(*result)
 }
 
 pub fn encode_string(input: &dyn AsRef<str>) -> String {
@@ -95,7 +36,11 @@ pub fn decode_string(input: &dyn AsRef<str>) -> Result<String, TranslationError>
     let input = input.as_ref();
     let result = {
         // Older versions used a ZWSP as a character separator, instead of `👉👈`.
-        if input.contains("\u{200B}") {
+        let split_char = input
+            .chars()
+            .find(|&c| c == '\u{200b}' || c == '👉');
+
+        if let Some('\u{200b}') = split_char {
             input.trim_end_matches("\u{200B}").split("\u{200B}")
         } else {
             input.trim_end_matches("👉👈").split("👉👈")
